Adds Task model, task routes and creating of order

+++ src/models/Item.kt
@@ -5,5 +5,5 @@ data class Item(
     val orderId: Int,
     val title: String,
     val info: String,
-    val price: Float
+    val price: Double
 )
