+++ src/repository/tables/Orders.kt
@@ -6,7 +6,7 @@ object Orders : Table() {
     val id = integer("id").autoIncrement()
     val customerName = varchar("customerName", 512)
     val customerEmail = varchar("customerEmail", 512)
-    val price = float("price")
+    val price = double("price")
     val createdAt = long("createdAt")
 
     override val primaryKey = PrimaryKey(id)
